+++ packages/shuttle/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @farcaster/hub-shuttle
 
+## 0.2.3
+
+### Patch Changes
+
+- feat(shuttle): add message state to handler
+
+## 0.2.2
+
+### Patch Changes
+
+- fix: ack the correct group name to fix stale bug
+
+## 0.2.1
+
+### Patch Changes
+
+- feat(shuttle): Make event stream more customizable
+
 ## 0.2.0
 
 ### Minor Changes